Working on mac

+++ .devcontainer/scripts/uninstall/macos.sh
@@ -1,13 +1,9 @@
 #!/usr/bin/env bash
 unsafe="$1"
-"$DEVCONTAINER_SCRIPTS_ROOT/utils/brew/uninstall.sh"
+"$DEVCONTAINER_SCRIPTS_ROOT/uninstall/brew.sh"
 sudo rm -rf "/Users/$(whoami)/Library/Caches/Homebrew/"
 sudo rm -rf "/Users/$(whoami)/Library/Logs/Homebrew/"
-sudo rm -rf /usr/local/Caskroom
-sudo rm -rf /usr/local/Cellar
-sudo rm -rf /usr/local/Homebrew
-case "$unsafe" in
- 1)
+if [ "$unsafe" -ge 1 ]; then
     sudo rm -rf /usr/local/Frameworks
     sudo rm -rf /usr/local/bin
     sudo rm -rf /usr/local/etc
@@ -17,8 +13,8 @@ case "$unsafe" in
     sudo rm -rf /usr/local/sbin
     sudo rm -rf /usr/local/share
     sudo rm -rf /usr/local/var
-    ;;
-  2)
+fi
+
+if [ "$unsafe" -ge 2 ]; then
     sudo rm -rf /usr/local/* &> /dev/null || true
-    ;;
-esac
+fi
